Refactor, splitting env_base.py into env_basic.py + env_manager.py
