+++ dcicutils/ff_mocks.py
@@ -17,7 +17,7 @@ from dcicutils.qa_utils import (
 )
 from dcicutils.s3_utils import EnvManager
 from unittest import mock
-from . import beanstalk_utils, ff_utils, s3_utils, env_utils, env_base  # , base
+from . import beanstalk_utils, ff_utils, s3_utils, env_utils, env_basic, env_manager  # env_base, base
 from .common import LEGACY_GLOBAL_ENV_BUCKET
 
 
@@ -135,37 +135,38 @@ def mocked_s3utils(environments=None, require_sse=False, other_access_key_names=
         with mock.patch.object(ff_utils, "boto3", mock_boto3):
             with mock.patch.object(beanstalk_utils, "boto3", mock_boto3):
                 with mock.patch.object(env_utils, "boto3", mock_boto3):
-                    with mock.patch.object(env_base, "boto3", mock_boto3):
-                        with mock.patch.object(s3_utils.EnvManager, "fetch_health_page_json") as mock_fetcher:
-                            # This is all that's needed for s3Utils to initialize an EnvManager.
-                            # We might have to add more later.
-                            def mocked_fetch_health_page_json(url, use_urllib=True):
-                                ignored(use_urllib)  # we don't test this
-                                m = re.match(r'.*(fourfront-[a-z0-9-]+)(?:[.]|$)', url)
-                                if m:
-                                    env_name = m.group(1)
-                                    return make_mock_health_page(env_name)
-                                else:
-                                    raise NotImplementedError(f"Mock can't handle URL: {url}")
+                    with mock.patch.object(env_basic, "boto3", mock_boto3):
+                        with mock.patch.object(env_manager, "boto3", mock_boto3):
+                            with mock.patch.object(s3_utils.EnvManager, "fetch_health_page_json") as mock_fetcher:
+                                # This is all that's needed for s3Utils to initialize an EnvManager.
+                                # We might have to add more later.
+                                def mocked_fetch_health_page_json(url, use_urllib=True):
+                                    ignored(use_urllib)  # we don't test this
+                                    m = re.match(r'.*(fourfront-[a-z0-9-]+)(?:[.]|$)', url)
+                                    if m:
+                                        env_name = m.group(1)
+                                        return make_mock_health_page(env_name)
+                                    else:
+                                        raise NotImplementedError(f"Mock can't handle URL: {url}")
 
-                            mock_fetcher.side_effect = mocked_fetch_health_page_json
-                            # The mocked encrypt key is expected by various tools in the s3_utils module
-                            # to be supplied as an environment variable (i.e., in os.environ), so this
-                            # sets up that environment variable.
-                            if require_sse:
-                                with override_environ(S3_ENCRYPT_KEY=s3_class.SSE_ENCRYPT_KEY):
+                                mock_fetcher.side_effect = mocked_fetch_health_page_json
+                                # The mocked encrypt key is expected by various tools in the s3_utils module
+                                # to be supplied as an environment variable (i.e., in os.environ), so this
+                                # sets up that environment variable.
+                                if require_sse:
+                                    with override_environ(S3_ENCRYPT_KEY=s3_class.SSE_ENCRYPT_KEY):
+                                        with EnvUtils.local_env_utils_for_testing(
+                                                global_env_bucket=os.environ.get('GLOBAL_ENV_BUCKET'),
+                                                env_name=(
+                                                        environments[0]
+                                                        if environments else
+                                                        os.environ.get('ENV_NAME'))):
+                                            yield mock_boto3
+                                else:
                                     with EnvUtils.local_env_utils_for_testing(
                                             global_env_bucket=os.environ.get('GLOBAL_ENV_BUCKET'),
-                                            env_name=(
-                                                    environments[0]
-                                                    if environments else
-                                                    os.environ.get('ENV_NAME'))):
+                                            env_name=os.environ.get('ENV_NAME')):
                                         yield mock_boto3
-                            else:
-                                with EnvUtils.local_env_utils_for_testing(
-                                        global_env_bucket=os.environ.get('GLOBAL_ENV_BUCKET'),
-                                        env_name=os.environ.get('ENV_NAME')):
-                                    yield mock_boto3
 
 
 # Here we set up some variables, auxiliary functions, and mocks containing common values needed for testing
