+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "dcicutils"
-version = "3.14.0.2b38"  # originally based on 3.6.0, but merges 3.14.0, eventually to be "4.0.0"
+version = "3.14.0.2b39"  # originally based on 3.6.0, but merges 3.14.0, eventually to be "4.0.0"
 description = "Utility package for interacting with the 4DN Data Portal and other 4DN resources"
 authors = ["4DN-DCIC Team <support@4dnucleome.org>"]
 license = "MIT"
